producer/requirements.txt
kafka-python==2.0.2
Faker==9.3.0
python-dotenv==1.0.0
pytest==6.2.5